# docs/swagger/traPortfolio.v1.yaml
openapi: 3.0.0
info:
  title: traPortfolio
  version: '1.0'
  contact:
    name: traP
    url: 'https://github.com/traPtitech/traPortfolio'
  description: traPortfolio v1 API
servers:
  - url: 'http://localhost:1323/api/v1'
    description: local
paths:
...
  /projects:
    get:
      summary: プロジェクトのリストを取得
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
      operationId: getProjects
      description: プロジェクトのリストを取得します
      tags:
        - project
    post:
      summary: プロジェクトの作成
      operationId: createProject
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetail'
        '400':
          description: Bad Request
      description: プロジェクトを作成します
      tags:
        - project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
  '/projects/{projectId}':
    parameters:
      - $ref: '#/components/parameters/projectIdInPath'
    get:
      summary: プロジェクト詳細情報を取得
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetail'
        '404':
          description: Not Found
      operationId: getProject
      description: プロジェクトの詳細情報を取得します
      tags:
        - project
    patch:
      summary: プロジェクト詳細情報の修正
      operationId: editProject
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
        '403':
          description: Forbidden
        '404':
          description: Not Found
      description: プロジェクト情報を修正します
      tags:
        - project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditProjectRequest'
...
  '/projects/{projectId}/members':
    parameters:
      - $ref: '#/components/parameters/projectIdInPath'
    get:
      summary: プロジェクトメンバーの取得
      tags:
        - project
        - user
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                description: プロジェクトメンバーの配列
                items:
                  $ref: '#/components/schemas/ProjectMember'
      operationId: getProjectMembers
      description: プロジェクトメンバーを取得します
    post:
      summary: プロジェクトメンバーの追加
      operationId: addProjectMembers
      description: プロジェクトメンバーを追加します
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
        '403':
          description: Forbidden
        '404':
          description: Not Found
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddProjectMembersRequest'
      tags:
        - project
        - user
    delete:
      summary: プロジェクトメンバーの削除
      operationId: deleteProjectMembers
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
        '403':
          description: Forbidden
        '404':
          description: Not Found
      description: プロジェクトメンバーを削除します
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberIDs'
      tags:
        - project
        - user
...
components:
  schemas:
    User:
      title: User
      type: object
      description: ユーザー情報
      properties:
        id:
          type: string
          description: ユーザーUUID
          format: uuid
          x-go-type: uuid.UUID
        name:
          type: string
          pattern: '^[a-zA-Z0-9_-]{1,32}$'
          description: ユーザー名
        realName:
          type: string
          description: 本名
      required:
        - id
        - name
        - realName
...
    Project:
      title: Project
      type: object
      description: プロジェクト情報
      properties:
        id:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          description: プロジェクトuuid
        name:
          type: string
          description: プロジェクト名
        duration:
          $ref: '#/components/schemas/YearWithSemesterDuration'
      required:
        - id
        - name
        - duration
...
    ProjectDetail:
      title: ProjectDetail
      type: object
      description: プロジェクトの詳細情報
      allOf:
        - $ref: '#/components/schemas/Project'
        - type: object
          properties:
            link:
              type: string
              format: uri
              description: プロジェクトの詳細が載っているページへのリンク
            description:
              type: string
              description: プロジェクト説明
            members:
              type: array
              description: プロジェクトメンバー
              items:
                $ref: '#/components/schemas/ProjectMember'
          required:
            - link
            - description
            - members
    ProjectMember:
      title: ProjectMember
      type: object
      description: プロジェクトメンバー情報
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            duration:
              $ref: '#/components/schemas/YearWithSemesterDuration'
          required:
            - duration
...
    YearWithSemesterDuration:
      title: YearWithSemesterDuration
      type: object
      properties:
        since:
          $ref: '#/components/schemas/YearWithSemester'
        until:
          $ref: '#/components/schemas/YearWithSemester'
      required:
        - since
      description: |-
        班やプロジェクトの期間
        年と前期/後期がある
        untilがなかった場合存続中
    YearWithSemester:
      title: YearWithSemester
      type: object
      properties:
        year:
          type: integer
        semester:
          $ref: '#/components/schemas/Semester'
      required:
        - year
        - semester
      description: 年度と前期/後期
    Semester:
      title: Semester
      type: integer
      description: |-
        0: 前期
        1: 後期
      enum:
        - 0
        - 1
      x-enum-varnames:
        - first
        - second
      x-enum-descriptions:
        - 前期
        - 後期
      format: int32
    MemberIDWithYearWithSemesterDuration:
      title: MemberIDWithYearWithSemesterDuration
      type: object
      description: プロジェクトメンバーのユーザーUUID(期間含む)
      properties:
        userId:
          type: string
          format: uuid
          x-go-type: uuid.UUID
        duration:
          $ref: '#/components/schemas/YearWithSemesterDuration'
      required:
        - userId
        - duration
    MemberIDs:
      title: MemberIDs
      type: object
      description: ユーザーのUUIDの配列
      properties:
        members:
          type: array
          description: ユーザーのUUIDの配列
          items:
            type: string
            format: uuid
            x-go-type: uuid.UUID
      required:
        - members
...
    CreateProjectRequest:
      title: CreateProjectRequest
      type: object
      description: 新規プロジェクトリクエスト
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 30
          description: プロジェクト名
        link:
          type: string
          format: uri
          description: プロジェクトの詳細が載っているページへのリンク
        description:
          type: string
          description: プロジェクト説明
        duration:
          $ref: '#/components/schemas/YearWithSemesterDuration'
      required:
        - name
        - description
        - duration
    EditProjectRequest:
      title: EditProjectRequest
      type: object
      description: プロジェクト変更リクエスト
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 30
          description: プロジェクト名
        link:
          type: string
          format: uri
          description: プロジェクトの詳細が載っているページへのリンク
        description:
          type: string
          description: プロジェクト説明
        duration:
          $ref: '#/components/schemas/YearWithSemesterDuration'
    AddProjectMembersRequest:
      title: AddProjectMembersRequest
      type: object
      description: プロジェクトメンバー追加リクエスト
      properties:
        members:
          type: array
          items:
            $ref: '#/components/schemas/MemberIDWithYearWithSemesterDuration'
      required:
        - members
...
  parameters:
...
    projectIdInPath:
      name: projectId
      in: path
      required: true
      description: プロジェクトUUID
      schema:
        type: string
        format: uuid
        x-go-type: uuid.UUID
...